Finished first pass of integration tests for Stories. Next would be the StoriesReference and then ProjectMetadata

+++ Taskter/ProjectAccess/Repositories/ProjectsAccess.cs
@@ -102,6 +102,9 @@ namespace ProjectsAccessComponent
                 //var project = projectsCollection.FindOne(projectToFind => projectToFind.ProjectAcronym == projectAcronym);
                 var project = projectsCollection.FindOne(Query.EQ("ProjectAcronym", projectAcronym));
 
+                if (project == null)
+                    return false;
+
                 return projectsCollection.Delete(project.Id);
             }
         }
